Update custom.css
wrapping answer option images for adaptive layout. It is different than the original IDS-2 layout for e.g. 5 images, but it is much more consistent imho. We can change the max-width parameter in the fw-adaptive script depending on how many images there are and how many images should be next to each other.